requirements.txt
numpy
sounddevice
scipy
openai-whisper
transformers
torch
edge-tts
pygame
